REFACTOR: Signup Refactor
