--- src/components/signup/SignupModal.tsx
@@ -1,24 +1,0 @@
-import React from 'react'
-import { Modal } from "react-bootstrap";
-import { SignUpError} from "../../hooks/Errors";
-
-interface Props {
-  setErrorShow: React.Dispatch<React.SetStateAction<boolean>>;
-  errorShow: boolean;
-  errorCode: string;
-}
-export default function SignupModal({setErrorShow,errorShow,errorCode}:Props) {
-     const handleClose = (e: React.MouseEvent<HTMLButtonElement> | void) => {
-       setErrorShow(false);
-     };
-  return (
-    <Modal show={errorShow} onHide={handleClose}>
-      <Modal.Body>{SignUpError(errorCode)}</Modal.Body>
-      <Modal.Footer>
-        <button className="btn btn-secondary" onClick={handleClose}>
-          닫기
-        </button>
-      </Modal.Footer>
-    </Modal>
-  );
-}
